+++ src/main/java/com/emailage/javawrapper/model/exception/EmailageApiRequestException.java
@@ -1,4 +1,4 @@
-package com.emailage.javawrapper.model;
+package com.emailage.javawrapper.model.exception;
 
 public class EmailageApiRequestException extends Exception {
     public EmailageApiRequestException(String message, Exception e){super(message,e);}
