+++ src/main/java/com/emailage/javawrapper/model/exception/EmailageParameterException.java
@@ -1,4 +1,4 @@
-package com.emailage.javawrapper.model;
+package com.emailage.javawrapper.model.exception;
 
 public class EmailageParameterException extends Exception {
     public EmailageParameterException(String message, Exception ex){super(message,ex);}
